+++ lib/src/providers/interface/llm_provider.dart
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:cross_file/cross_file.dart';
 import 'package:flutter/foundation.dart';
 
 import 'attachments.dart';
@@ -75,3 +76,6 @@ typedef LlmStreamGenerator =
       String prompt, {
       required Iterable<Attachment> attachments,
     });
+
+/// A function that converts speech to text.
+typedef SpeechToTextConverter = Stream<String> Function(XFile file);